third_party/WebKit/LayoutTests/http/tests/security/no-indexeddb-from-sandbox-expected.txt

CONSOLE MESSAGE: line 1: PASS: indexedDB.open() threw a SECURITY_ERR!
